hot fix for payload admin page
